backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Martin <martinpfatrisch@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
Flask = "^2.1.1"
psycopg-binary = "^3.0.11"
black = "^22.3.0"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"
paho-mqtt = "^1.6.1"
numpy = "^1.22.3"
pandas = "^1.4.2"
scipy = "^1.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
